backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
Pillow
# ultralytics # This is now expected to be in the base image 